Bugfixes
Fixed a bug when changing theme and on the options page the radio button colors did not change.
Fixed a bug with the previous commit so when setting a default language the application froze, so now the number group separators are Culture setting specific.
Fixed a bug with the chart where the chart entries were in a wrong order.